update roles, db constraint

+++ backend/app/Services/OrganizationWorkspace/OrganizationService.php
@@ -3,10 +3,9 @@
 namespace App\Services\OrganizationWorkspace;
 
 use App\Enums\WorkspaceType;
-use App\Http\Requests\OrganizationWorkspace\CreateOrganizationRequest;
 use App\Models\Organization;
 use App\Models\Workspace;
-use Illuminate\Http\Request;
+
 
 class OrganizationService
 {
@@ -17,18 +16,18 @@ class OrganizationService
         return $orgs;
     }
 
-    public function get(Request $request)
+    public function get($id)
     {
-        $org = Organization::find($request->id);
+        $org = Organization::find($id);
         $org->workspaces = $org->workspaces()->get();
         return $org;
     }
 
-    public function create(CreateOrganizationRequest $request)
+    public function create($name)
     {
         try {
-            $org = Organization::create(['name' => $request->name]);
-            $wsp = Workspace::create(['name' => $request->name, 'type' => WorkspaceType::corporation]);
+            $org = Organization::create(['name' => $name]);
+            $wsp = Workspace::create(['name' => $name, 'type' => WorkspaceType::corporation]);
             $org->save();
             $org->workspaces()->save($wsp);
             return $org;
@@ -39,10 +38,24 @@ class OrganizationService
 
     public function delete($id)
     {
+        if($id == 1)
+            return ['Public organization cannot be deleted'];
+
         $org = Organization::find($id);
         if($org != null) {
             $org->delete();
             return ['deleted' => $org];
+        } else {
+            return ['Organization not exists'];
+        }
+    }
+
+    public function update($id, $name)
+    {
+        $org = Organization::find($id);
+        if($org != null) {
+            $org->update(['name' => $name]);
+            return ['updated' => $org];
         } else {
             return ['Organization not exists'];
         }
